+++ ui/drawer/src/DrawerScrim.tsx
@@ -1,43 +1,31 @@
 import * as React from "react";
-import { styled } from "@washingtonpost/wpds-theme";
 import { DrawerContext } from "./DrawerRoot";
+import { Scrim } from "@washingtonpost/wpds-scrim";
+import type * as WPDS from "@washingtonpost/wpds-theme";
 
 const NAME = "DrawerScrim";
 
-const StyledElement = styled("div", {
-  backgroundColor: "rgba(0, 0, 0, 0.25)",
-  position: "fixed",
-  inset: 0,
-  display: "none",
-  variants: {
-    open: {
-      true: {
-        display: "block",
-      },
-    },
-  },
-});
-
-type DrawerScrimProps = React.ComponentPropsWithRef<typeof StyledElement>;
+type DrawerScrimProps = {
+  /** Override CSS */
+  css?: WPDS.CSS;
+};
 
 export const DrawerScrim = React.forwardRef<HTMLDivElement, DrawerScrimProps>(
   ({ children, ...props }, ref) => {
     const { onOpenChange, zIndex, open, defaultOpen } =
       React.useContext(DrawerContext);
     return (
-      <StyledElement
+      <Scrim
         onClick={() => {
           onOpenChange(false);
         }}
         ref={ref}
-        css={{
-          zIndex: `calc(${zIndex} - 1)`,
-        }}
+        zIndex={`calc(${zIndex} - 1)`}
         open={open || defaultOpen}
         {...props}
       >
         {children}
-      </StyledElement>
+      </Scrim>
     );
   }
 );