+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.17.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.16.0...v0.17.0) (2022-09-13)
+
+
+### Bug Fixes
+
+* styled component & text change ([7f40d70](https://github.com/washingtonpost/wpds-ui-kit/commit/7f40d70b21fd631415511a1ba6f57ffe80eab72c))
+
+
+### Features
+
+* add scrim component ([#211](https://github.com/washingtonpost/wpds-ui-kit/issues/211)) ([2cbc8b6](https://github.com/washingtonpost/wpds-ui-kit/commit/2cbc8b6573961aff2f754a2184fcc451cb8c1258))
+* create tooltip ([#212](https://github.com/washingtonpost/wpds-ui-kit/issues/212)) ([1250769](https://github.com/washingtonpost/wpds-ui-kit/commit/12507691245dae165817586f171e65e52d9718e7))
+* **docs:** add VisuallyHidden 'Jump to content' elem ([b347d9c](https://github.com/washingtonpost/wpds-ui-kit/commit/b347d9c556676fbe3d76b6071c3c20ee1671a7eb))
+* improve docs props table ([#210](https://github.com/washingtonpost/wpds-ui-kit/issues/210)) ([da36c33](https://github.com/washingtonpost/wpds-ui-kit/commit/da36c334355ee2285dc38b0300438c073d8e35d3))
+
+
+
+
+
 # [0.16.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.15.0...v0.16.0) (2022-09-02)