+++ ui/icon/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.17.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.16.0...v0.17.0) (2022-09-13)
+
+
+### Features
+
+* improve docs props table ([#210](https://github.com/washingtonpost/wpds-ui-kit/issues/210)) ([da36c33](https://github.com/washingtonpost/wpds-ui-kit/commit/da36c334355ee2285dc38b0300438c073d8e35d3))
+
+
+
+
+
 # [0.16.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.15.0...v0.16.0) (2022-09-02)
 
 **Note:** Version bump only for package @washingtonpost/wpds-icon